The Fork!
This update moves this fork of AMOLED-Cord away from its original and adds additional features to the theme, however the theme colour wise is almost identical to the original.
I have also fixed a couple references to the original fork as pointed out by @heikadog

Check the readme for the full list on changes in this fork!